+++ tests/run-all.sh
@@ -58,7 +58,7 @@ verify_completion_signal_classes \
 verify_task_contract_advisory \
 verify_review_evalset \
 verify_otel_export \
-verify_platform_tmp_and_new_hooks verify_goal_tools verify_goal_bounded_autonomy verify_retro_scan \
+verify_platform_tmp_and_new_hooks verify_goal_tools verify_goal_bounded_autonomy verify_goal_five_star verify_retro_scan \
 verify_v147_fixes verify_hook_depth verify_narration_final \
 verify_verdict_contract verify_worktree_hook_safety verify_hook_count_words \
 verify_fable_snippets verify_routing verify_completion_gate \